requirements.txt
flask==2.3.3
SpeechRecognition==3.10.0
pydub==0.25.1
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.3
librosa==0.10.1
soundfile==0.12.1
ffmpeg-python==0.2.0
requests==2.31.0
werkzeug==2.3.7
jinja2==3.1.2
